requirements.txt
Streamlitpip
langchain